refactor(config_routes): extract evolution.py + infrastructure.py + workflows.py from _factory.py (7/8)
Schritt 7 von 8 im Split-Plan
(`docs/superpowers/plans/2026-04-29-config-routes-split.md`).

What moved
- `_register_prompt_evolution_routes()` (~83 LOC) -> `cognithor.channels.config_routes.evolution`. Prompt-Evolution Toggle/Stats/Evolve.
- `_register_self_improvement_routes()` (~64 LOC) -> selbe Datei. Reflexion-getriebene Self-Improvement.
- `_register_gepa_evolution_routes()` (~141 LOC) -> selbe Datei. GEPA Evolution Loop. Inkl. `_proposal_to_dict` und `_trace_to_dict` Modul-Level-Helfer (zogen mit, weil exklusiv von gepa genutzt).
- `_register_infrastructure_routes()` (~80 LOC) -> `cognithor.channels.config_routes.infrastructure`. Ecosystem, Performance.
- `_register_portal_routes()` (~28 LOC) -> selbe Datei. End-User-Portal.
- `_register_backend_routes()` (~133 LOC) -> selbe Datei. LLM-Backend-Status / Switch.
- `_register_workflow_graph_routes()` (~169 LOC) -> `cognithor.channels.config_routes.workflows`. DAG-Workflow-Visualisierung.

`_factory.py` schrumpft von 1 952 -> 1 247 Zeilen (-36 %).

Per-Sub-Module-Imports
- `evolution.py`: `contextlib`, `Request`, `HTTPException`, `Any`, `log`. F401-clean.
- `infrastructure.py`: `asyncio`, `contextlib`, `Request`, `HTTPException`, `Any`, `ConfigManager` (TYPE_CHECKING). F401-clean.
- `workflows.py`: `json`, `Request`, `HTTPException`, `Any`, `log`. F401-clean.

`_factory.py` Cleanup
- `import asyncio` Top-Level entfernt — wandert mit `_register_backend_routes` nach infrastructure.py.
- `import json` Top-Level entfernt — wandert mit `_register_workflow_graph_routes` nach workflows.py.
- `import contextlib` bleibt (wird noch von `_register_learning_routes` benoetigt).

Aufruf-Reihenfolge in `create_config_routes()` exakt wie zuvor.

Verification
- `pytest test_route_inventory_unchanged` -> passed (303/303 endpoints).
- Konsumenten-Sweep ueber 11 Test-Files -> 396 passed.
- `ruff check` -> All checks passed.
- `ruff format --check` -> 13 files unchanged.
