+++ src/cognithor/channels/config_routes/_factory.py
@@ -20,9 +20,7 @@ Architektur-Bibel: §12 (Konfiguration), §9.3 (Web UI)
 
 from __future__ import annotations
 
-import asyncio
 import contextlib
-import json
 from typing import TYPE_CHECKING, Any
 
 try:
@@ -47,7 +45,17 @@ log = get_logger(__name__)
 
 
 from cognithor.channels.config_routes.config import _register_config_routes
+from cognithor.channels.config_routes.evolution import (
+    _register_gepa_evolution_routes,
+    _register_prompt_evolution_routes,
+    _register_self_improvement_routes,
+)
 from cognithor.channels.config_routes.governance import _register_governance_routes
+from cognithor.channels.config_routes.infrastructure import (
+    _register_backend_routes,
+    _register_infrastructure_routes,
+    _register_portal_routes,
+)
 from cognithor.channels.config_routes.monitoring import (
     _register_monitoring_routes,
     _register_prometheus_routes,
@@ -64,6 +72,7 @@ from cognithor.channels.config_routes.skills import (
 )
 from cognithor.channels.config_routes.system import _register_system_routes
 from cognithor.channels.config_routes.ui import _register_ui_routes
+from cognithor.channels.config_routes.workflows import _register_workflow_graph_routes
 
 # ======================================================================
 # Public entry-point
@@ -122,368 +131,6 @@ def create_config_routes(
     _register_autonomous_routes(app, deps, gateway)
     _register_feedback_routes(app, deps, gateway)
     _register_social_routes(app, deps, gateway)
-
-
-# ======================================================================
-# Prompt-Evolution routes
-# ======================================================================
-
-
-def _register_prompt_evolution_routes(
-    app: Any,
-    deps: list[Any],
-    gateway: Any,
-) -> None:
-    """Stats, manual evolve trigger, and enable/disable toggle."""
-
-    @app.get("/api/v1/prompt-evolution/stats", dependencies=deps)
-    async def prompt_evolution_stats() -> dict[str, Any]:
-        engine = getattr(gateway, "_prompt_evolution", None)
-        enabled = engine is not None
-        stats: dict[str, Any] = {"enabled": enabled}
-        if engine:
-            with contextlib.suppress(Exception):
-                stats.update(engine.get_stats("system_prompt"))
-        return stats
-
-    @app.post("/api/v1/prompt-evolution/evolve", dependencies=deps)
-    async def prompt_evolution_evolve() -> dict[str, Any]:
-        engine = getattr(gateway, "_prompt_evolution", None)
-        if engine is None:
-            return {"error": "prompt_evolution is disabled"}
-        # Check ImprovementGate
-        gate = getattr(gateway, "_improvement_gate", None)
-        if gate is not None:
-            from cognithor.governance.improvement_gate import GateVerdict, ImprovementDomain
-
-            verdict = gate.check(ImprovementDomain.PROMPT_TUNING)
-            if verdict != GateVerdict.ALLOWED:
-                return {"error": f"gate_blocked: {verdict.value}"}
-        try:
-            result = await engine.maybe_evolve("system_prompt")
-            return {"evolved": result is not None, "version_id": result}
-        except Exception as exc:
-            log.error("prompt_evolution_evolve_failed", error=str(exc))
-            return {"error": "Prompt-Evolution fehlgeschlagen"}
-
-    @app.post("/api/v1/prompt-evolution/toggle", dependencies=deps)
-    async def prompt_evolution_toggle(request: Request) -> dict[str, Any]:
-        body = await request.json()
-        enabled = body.get("enabled", False)
-
-        if enabled:
-            if getattr(gateway, "_prompt_evolution", None) is None:
-                try:
-                    from cognithor.learning.prompt_evolution import PromptEvolutionEngine
-
-                    cfg = gateway._config
-                    pe_db = str(cfg.db_path.with_name("memory_prompt_evolution.db"))
-                    engine = PromptEvolutionEngine(
-                        db_path=pe_db,
-                        min_sessions_per_arm=cfg.prompt_evolution.min_sessions_per_arm,
-                        significance_threshold=cfg.prompt_evolution.significance_threshold,
-                        max_concurrent_tests=cfg.prompt_evolution.max_concurrent_tests,
-                    )
-                    engine.set_evolution_interval_hours(
-                        cfg.prompt_evolution.evolution_interval_hours
-                    )
-                    gateway._prompt_evolution = engine
-                    planner = getattr(gateway, "_planner", None)
-                    if planner:
-                        planner._prompt_evolution = engine
-                except Exception as exc:
-                    log.error("prompt_evolution_toggle_failed", error=str(exc))
-                    return {
-                        "error": "Prompt-Evolution konnte nicht aktiviert werden",
-                        "enabled": False,
-                    }
-        else:
-            # Disable: disconnect from planner but keep engine for stats
-            planner = getattr(gateway, "_planner", None)
-            if planner:
-                planner._prompt_evolution = None
-            gateway._prompt_evolution = None
-
-        return {"enabled": getattr(gateway, "_prompt_evolution", None) is not None}
-
-
-# ======================================================================
-# Infrastructure routes (ecosystem, performance, portal)
-# ======================================================================
-
-
-def _register_infrastructure_routes(
-    app: Any,
-    deps: list[Any],
-    gateway: Any,
-) -> None:
-    """Ecosystem control, performance manager."""
-
-    # -- Ecosystem-Kontrolle (Phase 28) -----------------------------------
-
-    @app.get("/api/v1/ecosystem/stats", dependencies=deps)
-    async def ecosystem_stats() -> dict[str, Any]:
-        """Ecosystem-Controller Statistiken."""
-        ctrl = getattr(gateway, "_ecosystem_controller", None)
-        if ctrl is None:
-            return {"curator": {}, "fraud": {}}
-        return ctrl.stats()
-
-    @app.get("/api/v1/ecosystem/curator", dependencies=deps)
-    async def ecosystem_curator() -> dict[str, Any]:
-        """Kuration-Statistiken."""
-        ctrl = getattr(gateway, "_ecosystem_controller", None)
-        if ctrl is None:
-            return {"total_reviews": 0}
-        return ctrl.curator.stats()
-
-    @app.get("/api/v1/ecosystem/fraud", dependencies=deps)
-    async def ecosystem_fraud() -> dict[str, Any]:
-        """Fraud-Detection Statistiken."""
-        ctrl = getattr(gateway, "_ecosystem_controller", None)
-        if ctrl is None:
-            return {"total_signals": 0}
-        return ctrl.fraud.stats()
-
-    @app.get("/api/v1/ecosystem/training", dependencies=deps)
-    async def ecosystem_training() -> dict[str, Any]:
-        """Security-Training Status."""
-        ctrl = getattr(gateway, "_ecosystem_controller", None)
-        if ctrl is None:
-            return {"total_modules": 0}
-        return ctrl.trainer.stats()
-
-    @app.get("/api/v1/ecosystem/trust", dependencies=deps)
-    async def ecosystem_trust() -> dict[str, Any]:
-        """Trust-Boundary Statistiken."""
-        ctrl = getattr(gateway, "_ecosystem_controller", None)
-        if ctrl is None:
-            return {"total_boundaries": 0}
-        return ctrl.trust.stats()
-
-    # -- Performance-Manager (Phase 37) -----------------------------------
-
-    @app.get("/api/v1/performance/health", dependencies=deps)
-    async def perf_health() -> dict[str, Any]:
-        """Performance Health-Status."""
-        pm = getattr(gateway, "_perf_manager", None)
-        if pm is None:
-            return {"vector_store": {"entries": 0}}
-        return pm.health()
-
-    @app.get("/api/v1/performance/latency", dependencies=deps)
-    async def perf_latency() -> dict[str, Any]:
-        """Latenz-Statistiken."""
-        pm = getattr(gateway, "_perf_manager", None)
-        if pm is None:
-            return {"total_samples": 0}
-        return pm.latency.stats()
-
-    @app.get("/api/v1/performance/resources", dependencies=deps)
-    async def perf_resources() -> dict[str, Any]:
-        """Ressourcen-Auslastung."""
-        pm = getattr(gateway, "_perf_manager", None)
-        if pm is None:
-            return {"snapshots": 0}
-        return pm.optimizer.stats()
-
-
-# ======================================================================
-# Portal routes (end-user portal)
-# ======================================================================
-
-
-def _register_portal_routes(
-    app: Any,
-    deps: list[Any],
-    gateway: Any,
-) -> None:
-    """End-user portal, consent management."""
-
-    @app.get("/api/v1/portal/stats", dependencies=deps)
-    async def portal_stats() -> dict[str, Any]:
-        """Endnutzer-Portal Statistiken."""
-        up = getattr(gateway, "_user_portal", None)
-        if up is None:
-            return {"consents": {"total_users": 0}}
-        return up.stats()
-
-    @app.get("/api/v1/portal/consents", dependencies=deps)
-    async def portal_consents() -> dict[str, Any]:
-        """Consent-Management Status."""
-        up = getattr(gateway, "_user_portal", None)
-        if up is None:
-            return {"total_users": 0}
-        return up.consents.stats()
-
-
-# ======================================================================
-# Workflow Execution Graph API
-# ======================================================================
-
-
-def _register_workflow_graph_routes(
-    app: Any,
-    deps: list[Any],
-    gateway: Any,
-) -> None:
-    """REST endpoints for workflow execution graph visualization."""
-
-    def _get_engines() -> tuple[Any, Any, Any]:
-        """Return (simple_engine, dag_engine, template_library) from gateway."""
-        simple = getattr(gateway, "_workflow_engine", None) if gateway else None
-        dag = getattr(gateway, "_dag_workflow_engine", None) if gateway else None
-        tmpl = getattr(gateway, "_template_library", None) if gateway else None
-        return simple, dag, tmpl
-
-    # -- Templates ---------------------------------------------------------
-
-    @app.get("/api/v1/workflows/templates", dependencies=deps)
-    async def wf_list_templates() -> dict[str, Any]:
-        """List all available workflow templates."""
-        _, _, tmpl = _get_engines()
-        if not tmpl:
-            return {"templates": [], "count": 0}
-        return {"templates": tmpl.list_all(), "count": tmpl.template_count}
-
-    @app.get("/api/v1/workflows/templates/{template_id}", dependencies=deps)
-    async def wf_get_template(template_id: str) -> dict[str, Any]:
-        _, _, tmpl = _get_engines()
-        if not tmpl:
-            return {"error": "Template library unavailable", "status": 503}
-        t = tmpl.get(template_id)
-        if not t:
-            return {"error": "Template not found", "status": 404}
-        return t.to_dict()
-
-    # -- Simple workflow instances -----------------------------------------
-
-    @app.get("/api/v1/workflows/instances", dependencies=deps)
-    async def wf_list_instances() -> dict[str, Any]:
-        """List all workflow instances (simple engine)."""
-        simple, _, _ = _get_engines()
-        if not simple:
-            return {"instances": [], "stats": {}}
-        all_inst = list(simple._instances.values())
-        return {
-            "instances": [i.to_dict() for i in all_inst],
-            "stats": simple.stats(),
-        }
-
-    @app.get("/api/v1/workflows/instances/{instance_id}", dependencies=deps)
-    async def wf_get_instance(instance_id: str) -> dict[str, Any]:
-        simple, _, tmpl = _get_engines()
-        if not simple:
-            return {"error": "Workflow engine unavailable", "status": 503}
-        inst = simple.get(instance_id)
-        if not inst:
-            return {"error": "Instance not found", "status": 404}
-        result = inst.to_dict()
-        result["step_results"] = inst.step_results
-        if tmpl:
-            t = tmpl.get(inst.template_id)
-            if t:
-                result["steps"] = [s.to_dict() for s in t.steps]
-        return result
-
-    @app.post("/api/v1/workflows/instances", dependencies=deps)
-    async def wf_start_instance(request: Request) -> dict[str, Any]:
-        """Start a new workflow from a template."""
-        simple, _, tmpl = _get_engines()
-        if not simple or not tmpl:
-            return {"error": "Workflow engine unavailable", "status": 503}
-        body = await request.json()
-        template_id = body.get("template_id", "")
-        t = tmpl.get(template_id)
-        if not t:
-            return {"error": f"Template '{template_id}' not found", "status": 404}
-        inst = simple.start(t, created_by=body.get("created_by", "ui"))
-        return {"status": "ok", "instance": inst.to_dict()}
-
-    # -- DAG workflow runs -------------------------------------------------
-
-    @app.get("/api/v1/workflows/dag/runs", dependencies=deps)
-    async def wf_list_dag_runs() -> dict[str, Any]:
-        """List DAG workflow runs (checkpoint-based)."""
-        _, dag, _ = _get_engines()
-        if not dag or not dag._checkpoint_dir:
-            return {"runs": []}
-        cp_dir = dag._checkpoint_dir
-        runs = []
-        if cp_dir.exists():
-            for cp_file in sorted(cp_dir.glob("*.json"), reverse=True):
-                try:
-                    data = json.loads(cp_file.read_text(encoding="utf-8"))
-                    runs.append(
-                        {
-                            "id": data.get("id", ""),
-                            "workflow_id": data.get("workflow_id", ""),
-                            "workflow_name": data.get("workflow_name", ""),
-                            "status": data.get("status", ""),
-                            "started_at": data.get("started_at"),
-                            "completed_at": data.get("completed_at"),
-                            "node_count": len(data.get("node_results", {})),
-                        }
-                    )
-                except Exception:
-                    continue
-        return {"runs": runs}
-
-    @app.get("/api/v1/workflows/dag/runs/{run_id}", dependencies=deps)
-    async def wf_get_dag_run(run_id: str) -> dict[str, Any]:
-        """Get full DAG workflow run with node graph data."""
-        _, dag, _ = _get_engines()
-        if not dag or not dag._checkpoint_dir:
-            return {"error": "DAG engine unavailable", "status": 503}
-        cp_file = (dag._checkpoint_dir / f"{run_id}.json").resolve()
-        try:
-            cp_file.relative_to(dag._checkpoint_dir.resolve())
-        except ValueError:
-            return {"error": "Invalid run_id (Path-Traversal)", "status": 400}
-        if not cp_file.exists():
-            return {"error": "Run not found", "status": 404}
-        try:
-            return json.loads(cp_file.read_text(encoding="utf-8"))
-        except Exception as exc:
-            log.error("wf_dag_run_read_failed", run_id=run_id, error=str(exc))
-            return {"error": "DAG-Run konnte nicht geladen werden", "status": 500}
-
-    @app.get("/api/v1/workflows/dag/runs/{run_id}/nodes/{node_id}", dependencies=deps)
-    async def wf_get_dag_node_detail(run_id: str, node_id: str) -> dict[str, Any]:
-        """Get detailed execution data for a single DAG node."""
-        _, dag, _ = _get_engines()
-        if not dag or not dag._checkpoint_dir:
-            return {"error": "DAG engine unavailable", "status": 503}
-        cp_file = (dag._checkpoint_dir / f"{run_id}.json").resolve()
-        try:
-            cp_file.relative_to(dag._checkpoint_dir.resolve())
-        except ValueError:
-            return {"error": "Invalid run_id (Path-Traversal)", "status": 400}
-        if not cp_file.exists():
-            return {"error": "Run not found", "status": 404}
-        try:
-            data = json.loads(cp_file.read_text(encoding="utf-8"))
-            node_results = data.get("node_results", {})
-            if node_id not in node_results:
-                return {"error": f"Node '{node_id}' not found in run", "status": 404}
-            return {"node_id": node_id, "run_id": run_id, **node_results[node_id]}
-        except json.JSONDecodeError:
-            return {"error": "Invalid run data", "status": 500}
-
-    # -- Combined stats ----------------------------------------------------
-
-    @app.get("/api/v1/workflows/stats", dependencies=deps)
-    async def wf_stats() -> dict[str, Any]:
-        """Combined workflow stats."""
-        simple, dag, tmpl = _get_engines()
-        result: dict[str, Any] = {"templates": 0, "simple": {}, "dag_runs": 0}
-        if tmpl:
-            result["templates"] = tmpl.template_count
-        if simple:
-            result["simple"] = simple.stats()
-        if dag and dag._checkpoint_dir and dag._checkpoint_dir.exists():
-            result["dag_runs"] = len(list(dag._checkpoint_dir.glob("*.json")))
-        return result
 
 
 # ======================================================================
@@ -1120,344 +767,6 @@ def _register_ingest_routes(
             return {"error": "Knowledge ingest service not initialized", "status": 503}
 
         return svc.stats()
-
-
-# ======================================================================
-# Self-improvement routes
-# ======================================================================
-
-
-def _register_self_improvement_routes(
-    app: Any,
-    deps: list[Any],
-    gateway: Any,
-) -> None:
-    """REST endpoints for the self-improvement engine."""
-
-    def _get_improver() -> Any:
-        return getattr(gateway, "_self_improver", None) if gateway else None
-
-    @app.get("/api/v1/learning/self-improvement/stats", dependencies=deps)
-    async def self_improvement_stats() -> dict[str, Any]:
-        """Return self-improvement engine statistics."""
-        improver = _get_improver()
-        if not improver:
-            return {"error": "Self-improvement engine not initialized", "status": 503}
-        return improver.stats()
-
-    @app.get("/api/v1/learning/self-improvement/pending", dependencies=deps)
-    async def self_improvement_pending() -> dict[str, Any]:
-        """Return pending improvement proposals."""
-        improver = _get_improver()
-        if not improver:
-            return {"error": "Self-improvement engine not initialized", "status": 503}
-
-        pending = improver.pending_improvements
-        return {
-            "pending": [
-                {
-                    "id": imp.id,
-                    "pattern_id": imp.pattern_id,
-                    "improvement_type": imp.improvement_type,
-                    "before": imp.before,
-                    "after": imp.after,
-                    "confidence": round(imp.confidence, 3),
-                    "created_at": imp.created_at.isoformat(),
-                }
-                for imp in pending
-            ],
-            "count": len(pending),
-        }
-
-    @app.post(
-        "/api/v1/learning/self-improvement/{improvement_id}/apply",
-        dependencies=deps,
-    )
-    async def self_improvement_apply(improvement_id: str) -> dict[str, Any]:
-        """Apply a pending improvement."""
-        improver = _get_improver()
-        if not improver:
-            return {"error": "Self-improvement engine not initialized", "status": 503}
-
-        success = improver.apply_improvement(improvement_id)
-        if not success:
-            return {"error": f"Improvement '{improvement_id}' not found", "status": 404}
-
-        return {"applied": True, "improvement_id": improvement_id}
-
-
-# ======================================================================
-# GEPA Evolution routes
-# ======================================================================
-
-
-def _register_gepa_evolution_routes(
-    app: Any,
-    deps: list[Any],
-    gateway: Any,
-) -> None:
-    """REST endpoints for GEPA (Guided Evolution through Pattern Analysis)."""
-
-    def _get_orch() -> Any:
-        return getattr(gateway, "_evolution_orchestrator", None) if gateway else None
-
-    def _get_trace_store() -> Any:
-        return getattr(gateway, "_trace_store", None) if gateway else None
-
-    def _get_proposal_store() -> Any:
-        return getattr(gateway, "_proposal_store", None) if gateway else None
-
-    @app.get("/api/v1/evolution/status", dependencies=deps)
-    async def get_evolution_status() -> dict[str, Any]:
-        orch = _get_orch()
-        if not orch:
-            return {"enabled": False, "message": "GEPA not enabled"}
-        return orch.get_status()
-
-    @app.get("/api/v1/learning/gepa/status", dependencies=deps)
-    async def gepa_status() -> dict[str, Any]:
-        """Get GEPA evolution cycle status (alias under /learning/)."""
-        orch = _get_orch()
-        if orch is None:
-            return {"status": "not_initialized"}
-        try:
-            status = orch.get_status()
-            return {"status": "ok", **status}
-        except Exception as exc:
-            return {"status": "error", "error": str(exc)}
-
-    @app.get("/api/v1/evolution/proposals", dependencies=deps)
-    async def list_evolution_proposals(status: str = "all") -> dict[str, Any]:
-        ps = _get_proposal_store()
-        if not ps:
-            return {"proposals": []}
-        if status == "all":
-            proposals = ps.get_history(limit=50)
-        else:
-            proposals = ps.get_by_status(status)
-        return {"proposals": [_proposal_to_dict(p) for p in proposals]}
-
-    @app.get("/api/v1/evolution/proposals/{proposal_id}", dependencies=deps)
-    async def get_evolution_proposal(proposal_id: str) -> dict[str, Any]:
-        orch = _get_orch()
-        if not orch:
-            return {"error": "GEPA not enabled", "status": 404}
-        detail = orch.get_proposal_detail(proposal_id)
-        if not detail:
-            return {"error": "Proposal not found", "status": 404}
-        return detail
-
-    @app.post("/api/v1/evolution/proposals/{proposal_id}/apply", dependencies=deps)
-    async def apply_evolution_proposal(proposal_id: str) -> dict[str, Any]:
-        orch = _get_orch()
-        if not orch:
-            return {"error": "GEPA not enabled", "status": 404}
-        ok = orch.apply_proposal(proposal_id)
-        return {"applied": ok, "proposal_id": proposal_id}
-
-    @app.post("/api/v1/evolution/proposals/{proposal_id}/reject", dependencies=deps)
-    async def reject_evolution_proposal(proposal_id: str) -> dict[str, Any]:
-        orch = _get_orch()
-        if not orch:
-            return {"error": "GEPA not enabled", "status": 404}
-        ok = orch.reject_proposal(proposal_id)
-        return {"rejected": ok, "proposal_id": proposal_id}
-
-    @app.post("/api/v1/evolution/proposals/{proposal_id}/rollback", dependencies=deps)
-    async def rollback_evolution_proposal(proposal_id: str) -> dict[str, Any]:
-        orch = _get_orch()
-        if not orch:
-            return {"error": "GEPA not enabled", "status": 404}
-        ok = orch.rollback_proposal(proposal_id)
-        return {"rolled_back": ok, "proposal_id": proposal_id}
-
-    @app.get("/api/v1/evolution/traces", dependencies=deps)
-    async def list_evolution_traces(limit: int = 20) -> dict[str, Any]:
-        ts = _get_trace_store()
-        if not ts:
-            return {"traces": []}
-        traces = ts.get_recent_traces(limit=min(limit, 100))
-        return {"traces": [_trace_to_dict(t) for t in traces]}
-
-    @app.post("/api/v1/evolution/run", dependencies=deps)
-    async def trigger_evolution_cycle() -> dict[str, Any]:
-        orch = _get_orch()
-        if not orch:
-            return {"error": "GEPA not enabled", "status": 404}
-        result = orch.run_evolution_cycle()
-        return {
-            "cycle_id": result.cycle_id,
-            "traces_analyzed": result.traces_analyzed,
-            "findings": result.findings_count,
-            "proposals_generated": result.proposals_generated,
-            "proposal_applied": result.proposal_applied,
-            "auto_rollbacks": result.auto_rollbacks,
-            "duration_ms": result.duration_ms,
-        }
-
-
-def _proposal_to_dict(p: Any) -> dict[str, Any]:
-    return {
-        "proposal_id": p.proposal_id,
-        "optimization_type": p.optimization_type,
-        "target": p.target,
-        "description": p.description,
-        "confidence": p.confidence,
-        "estimated_impact": p.estimated_impact,
-        "failure_category": p.failure_category,
-        "tool_name": p.tool_name,
-        "status": p.status,
-        "created_at": p.created_at,
-        "applied_at": p.applied_at,
-    }
-
-
-def _trace_to_dict(t: Any) -> dict[str, Any]:
-    return {
-        "trace_id": t.trace_id,
-        "session_id": t.session_id,
-        "goal": t.goal[:200],
-        "success_score": t.success_score,
-        "model_used": t.model_used,
-        "total_duration_ms": t.total_duration_ms,
-        "step_count": len(t.steps),
-        "failed_steps": len(t.failed_steps),
-        "tool_sequence": t.tool_sequence,
-        "created_at": t.created_at,
-    }
-
-
-# ======================================================================
-# Backend status / switch routes
-# ======================================================================
-
-
-def _register_backend_routes(
-    app: Any,
-    deps: list[Any],
-    config_manager: ConfigManager,
-    gateway: Any,
-) -> None:
-    """Endpoints for querying LLM backend availability and switching backends."""
-
-    import shutil
-
-    @app.get("/api/v1/backend/status", dependencies=deps)
-    async def get_backend_status() -> dict[str, Any]:
-        """Check which LLM backends are available and authenticated."""
-        results: dict[str, Any] = {}
-
-        # Claude Code CLI
-        claude_path = shutil.which("claude")
-        results["claude-code"] = {
-            "installed": claude_path is not None,
-            "path": claude_path or "",
-            "authenticated": False,
-            "models": [],
-        }
-        if claude_path:
-            try:
-                proc = await asyncio.create_subprocess_exec(
-                    claude_path,
-                    "--version",
-                    stdout=asyncio.subprocess.PIPE,
-                    stderr=asyncio.subprocess.PIPE,
-                )
-                stdout, _ = await asyncio.wait_for(proc.communicate(), timeout=5)
-                if proc.returncode == 0:
-                    results["claude-code"]["authenticated"] = True
-                    results["claude-code"]["version"] = stdout.decode().strip()
-                    results["claude-code"]["models"] = [
-                        "opus",
-                        "sonnet",
-                        "haiku",
-                    ]
-            except Exception:
-                log.debug("claude_code_provider_check_failed", exc_info=True)
-
-        # Ollama
-        try:
-            import httpx
-
-            async with httpx.AsyncClient() as client:
-                resp = await client.get("http://localhost:11434/api/tags", timeout=3)
-                if resp.status_code == 200:
-                    models = [m["name"] for m in resp.json().get("models", [])]
-                    results["ollama"] = {
-                        "installed": True,
-                        "authenticated": True,
-                        "models": models,
-                    }
-                else:
-                    results["ollama"] = {
-                        "installed": True,
-                        "authenticated": False,
-                        "models": [],
-                    }
-        except Exception:
-            results["ollama"] = {
-                "installed": False,
-                "authenticated": False,
-                "models": [],
-            }
-
-        # OpenAI (check if key is set)
-        cfg = config_manager.config if config_manager else None
-        if cfg:
-            has_openai = bool(getattr(cfg, "openai_api_key", ""))
-            results["openai"] = {
-                "installed": True,
-                "authenticated": has_openai,
-                "models": [],
-            }
-
-            has_anthropic = bool(getattr(cfg, "anthropic_api_key", ""))
-            results["anthropic"] = {
-                "installed": True,
-                "authenticated": has_anthropic,
-                "models": [],
-            }
-
-            has_openrouter = bool(getattr(cfg, "openrouter_api_key", ""))
-            results["openrouter"] = {
-                "installed": True,
-                "authenticated": has_openrouter,
-                "models": [],
-            }
-
-        # Current backend
-        current = getattr(cfg, "llm_backend_type", "ollama") if cfg else "ollama"
-
-        return {"backends": results, "current": current}
-
-    @app.post("/api/v1/backend/switch", dependencies=deps)
-    async def switch_backend(request: Request) -> dict[str, Any]:
-        """Switch the LLM backend type."""
-        body = await request.json()
-        new_backend = body.get("backend", "")
-
-        # Derive valid backends from the config Literal type (single source of truth)
-        from typing import get_args, get_type_hints
-
-        from cognithor.config import CognithorConfig
-
-        _hints = get_type_hints(CognithorConfig, include_extras=True)
-        valid = list(get_args(_hints["llm_backend_type"]))
-        if new_backend not in valid:
-            raise HTTPException(400, f"Invalid backend: {new_backend}. Valid: {valid}")
-
-        # Update config
-        if config_manager:
-            config_manager.config.llm_backend_type = new_backend
-            # Save to config.yaml
-            with contextlib.suppress(Exception):
-                config_manager.save()
-
-        return {
-            "status": "switched",
-            "backend": new_backend,
-            "note": "Restart required for full effect",
-        }
 
 
 # ======================================================================
